requirements.txt
appdirs==1.4.3
asgiref==3.4.1
astroid==2.7.2
beautifulsoup4==4.8.1
bootstrap4==0.1.0
CacheControl==0.12.6
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.2.6
django-appconf==1.0.3
django-autocomplete-light==3.8.2
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==3.0.1
django-countries==5.5
django-crispy-forms==1.12.0
django-cruds-adminlte==0.0.16
django-forms-bootstrap==3.1.0
django-image-cropping==1.2.0
django-modeltranslation==0.13.4
django-rosetta==0.9.3
django-settings-export==1.2.1
django-taggit==1.1.0
django-userforeignkey==0.3.0
djangoajax==3.1
djangorestframework==3.12.4
easy-thumbnails==2.6
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
isort==4.3.21
lazy-object-proxy==1.4.3
lockfile==0.12.2
mccabe==0.6.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
Pillow==8.3.1
pipenv==2018.11.26
pkg_resources==0.0.0
platformdirs==2.2.0
polib==1.1.0
progress==1.5
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pylint==2.10.2
pylint-django==2.0.11
pylint-plugin-utils==0.6
pyparsing==2.4.6
PyPDF2==1.26.0
pytoml==0.1.21
pytz==2019.3
reportlab==3.5.31
requests==2.22.0
retrying==1.3.3
six==1.16.0
soupsieve==1.9.4
sqlparse==0.3.0
toml==0.10.2
typed-ast==1.4.0
urllib3==1.25.3
virtualenv==16.7.7
virtualenv-clone==0.5.3
webencodings==0.5.1
wrapt==1.11.2
xhtml2pdf==0.2.3
